convert_mesh.py: new --extract-surface, --print-surface options in main()
- move _get_facets(), get_surface_faces(), surface_graph(),
  surface_components() into mesh_tools.py
- remove extract_surface.py
